requirements.txt
notebook
ipywidgets
tqdm
python-dotenv
chainlit
langchain
openai
pandasai